+++ TablePro/Views/Sidebar/SidebarView.swift
@@ -372,7 +372,8 @@ struct SidebarView: View {
             activeDatabase: activeDatabase,
             activeSchema: coordinator?.toolbarState.currentSchema,
             selectedTables: windowState.selectedTables,
-            showRecentTables: settingsManager.general.showRecentTables
+            showRecentTables: settingsManager.general.showRecentTables,
+            rowSizePreference: settingsManager.general.sidebarRowSize
         )
     }